parse nostr metadat
